Phase 7 Week 25: GPU Profiler Queries Infrastructure
- Implement VulkanQueryPool for Vulkan timestamp management
  * Query pool creation and lifecycle management
  * GPU→CPU result retrieval with synchronization
  * Circular buffering to avoid stalls
  * Timestamp to milliseconds conversion

- Create GPUProfiler for per-pass timing aggregation
  * Per-pass GPU timing measurement
  * 100-frame historical data storage
  * 10-frame rolling average calculation
  * Thread-safe singleton for global access

- Develop GPU Profiler UI Panels with ImGui
  * Real-time statistics table (per-pass timing)
  * Historical timeline visualization (line graphs)
  * GPU load percentage and budget tracking
  * Color-coded bottleneck highlighting

- Add render graph integration guide
  * Clear instructions for query placement
  * Timestamp synchronization patterns
  * Latency hiding strategy (2-frame delay)

- Update CMakeLists.txt with all new GPU profiler files

Total: 755 lines of GPU profiling infrastructure
Files: query_pool (240 LOC), gpu_profiler (230 LOC), gpu_profiler_panels (195 LOC)
Documentation: Comprehensive Week 25 guide with performance notes

Target: GPU timing <1% variance, <0.5% profiler overhead
Ready for: Render graph integration in Week 25 continuation

+++ engine/renderer/gpu/vulkan/vulkan_swapchain.cpp
@@ -1,5 +1,5 @@
 #include "vulkan_swapchain.h"
-#include "gws/core/logging/logger.h"
+#include "logging/logger.h"
 #include <algorithm>
 
 namespace gws::renderer::gpu {
@@ -153,7 +153,7 @@ uint32_t VulkanSwapchain::acquire_next_image(VkSemaphore signal_semaphore) {
             // Swapchain is out of date, needs recreation
             GWS_LOG_WARN("Swapchain out of date, needs recreation");
         } else {
-            GWS_LOG_ERROR("Failed to acquire next image: {}", result);
+            GWS_LOG_ERROR("Failed to acquire next image: {}", static_cast<int>(result));
         }
         return ~0u;
     }
@@ -177,7 +177,7 @@ void VulkanSwapchain::present_image(uint32_t image_index, VkSemaphore wait_semap
         if (result == VK_ERROR_OUT_OF_DATE_KHR) {
             GWS_LOG_WARN("Swapchain out of date during present");
         } else {
-            GWS_LOG_ERROR("Failed to present: {}", result);
+            GWS_LOG_ERROR("Failed to present: {}", static_cast<int>(result));
         }
     }
 }
